refactor(hindsight): split venues into clients and solvers modules
The venues directory mixed both tiers of the taxonomy with the
tier-neutral machinery. Client decoding (relay, metamask, fee back-out,
the name-to-strategy binding) now lives in clients/, solver calldata
knowledge (quote extraction, lifi's bridge veto) in solvers/, and the
neutral seam (Strategy, DecodeContext, Flow, select) in strategy.rs,
with intent maker-finding beside it. Under the project terminology,
'venues' now only ever means liquidity venues, which the decoder does
not model.
